Create quaternary simplex plot with pre-calculated similarity matrix

Create quaternary simplex plot with pre-calculated similarity matrix

Usage

# S3 method for class 'simMat'
plotQuaternary(
  x,
  veloMat = NULL,
  nGrid = 10,
  radius = 0.2,
  dotSize = NULL,
  colorArg = NULL,
  labelColors = c("#3B4992FF", "#EE0000FF", "#008B45FF", "#631879FF"),
  arrowLinewidth = NULL,
  arrowAngle = 20,
  arrowLen = 0.1,
  vertexLabelSize = NULL,
  edgeLinewidth = 1,
  title = NULL,
  titleSize = 1,
  titleColor = "black",
  theta = 20,
  phi = 0,
  interactive = TRUE,
  ...
)

Arguments

x

simMat object, n cells by 4 vertices, each row summing to 1.

veloMat

Aggregated velocity matrix. Output of aggrVeloGraph.

nGrid

Number of grids along the x-axis of the tetrahedron. Default 10.

radius

Arrow length of unit velocity. Lower this when arrows point outside of the tetrahedron. Default 0.2.

dotSize

Size of each dot. Default 0.6 for static figure, and 4 for interactive view.

colorArg

A "colorArg" object, internally prepared by plotQuaternary.default. Default NULL.

labelColors

Colors of the vertex labels. Default c("#3B4992FF", "#EE0000FF", "#008B45FF", "#631879FF") (blue, red, green and purple).

arrowLinewidth

Arrow aesthetics. Default 1.6 for interactive view, 0.6 for static figure.

arrowAngle, arrowLen

Arrow aesthetics passed to TODOOOO grid::arrow. The length of the arrow will be internally converted to unit onject in inches. Default 20 and 0.1.

vertexLabelSize

Numeric, size of vertex text label relative to default size. Default 1.

edgeLinewidth

Controls the linewidth of the edges of the tetrahedron. Default 1.

title

Title text of the plot. Default NULL.

titleSize, titleColor

Setting on the main title text. Default 1, and "black".

theta, phi

Numeric scalar. The angles defining the viewing direction. theta gives the azimuthal direction and phi the colatitude. Default 20 and 0.

interactive

Logical. Whether to display plotly interactive view. Default TRUE.

...

Not used

Value

A "quatPlot" object, can be displayed by printing.